split du filtre et de liste ds dressing
