+++ app/components/dressing/dressingList.component.tsx
@@ -1,17 +1,15 @@
 import { Ionicons } from "@expo/vector-icons";
 import { ThemedView } from "../commons/ThemedView";
 import { ThemedText } from "../commons/ThemedText";
-import { Pressable, View, Image } from "react-native";
+import { Pressable, Image } from "react-native";
 import VetementModel from "@/app/models/vetements.model";
 import { Colors } from "@/constants/Colors";
 import { VetemenItemComponent } from "./vetementItem.component";
-import { calculFiltresPossibles, groupeVetementByType, setVetementsFiltres as applyFiltresOnVetements, selectFilters as updateSelectedFilters } from "@/app/controllers/dressingList.controller";
-import { MultiSelect } from "react-native-element-dropdown";
-import { useEffect, useState } from "react";
-import DressingListFiltreModel from "@/app/models/dressingListeFiltre.model";
+import { groupeVetementByType } from "@/app/controllers/dressingList.controller";
+import { useState } from "react";
 import { alphanumSort, getTypeVetementIcon, vetementSort } from "../commons/CommonsUtils";
-import { CaracteristiqueVetementEnum, StatutVetementEnum } from "@/constants/AppEnum";
 import { styles } from "./dressingList.style";
+import { DressingFiltreComponent } from "./dressingFiltres.component";
 
 
 export type DressingComponentProps = {
@@ -29,60 +27,8 @@ export type DressingComponentProps = {
  **/
 export const DressingListComponent: React.FC<DressingComponentProps> = ({ vetementsInDressing, openAddEditVetement }: DressingComponentProps) => {
 
-
-    const [selectedFiltres, setSelectedFiltres] = useState<DressingListFiltreModel[]>([
-        {id: StatutVetementEnum.ACTIF, libelle: StatutVetementEnum.ACTIF, type: CaracteristiqueVetementEnum.STATUT, typeLibelle: CaracteristiqueVetementEnum.STATUT+StatutVetementEnum.ACTIF},]);
-    const [filtresDisponibles, setFiltresDisponibles] = useState<DressingListFiltreModel[]>([]);
-
     const [vetementsAffiches, setVetementsAffiches] = useState<VetementModel[]>([]);
 
-    useEffect(() => {
-        // Recalcul des filtres disponibles
-        setFiltresDisponibles(calculFiltresPossibles(vetementsInDressing));
-    }, [vetementsInDressing]);
-
-    useEffect(() => {
-        // Mise à jour de l'affichage des vêtements en fonction des filtres sélectionnés
-        setVetementsAffiches(applyFiltresOnVetements(vetementsInDressing, selectedFiltres));
-    }, [selectedFiltres]);
-
-
-
-    /**
-     * Rendu de la barre de filtres.
-     *
-     * Cette fonction retourne une vue thématisée contenant un composant MultiSelect
-     * permettant de sélectionner un ou plusieurs filtres parmi les filtres disponibles.
-     *
-     * @returns {JSX.Element} La vue thématisée avec le composant MultiSelect.
-     */
-    function renderFiltersBar() {
-        return (
-            
-            <ThemedView style={styles.filtresBar}>
-                <View style={styles.filtresBar}>
-                    <MultiSelect
-                        style={styles.dropdown} containerStyle={styles.listStyle} itemContainerStyle={styles.listItemStyle} itemTextStyle={styles.listItemStyle}
-                        iconStyle={styles.iconStyle} activeColor={Colors.app.color} placeholderStyle={styles.placeholderStyle} selectedTextStyle={styles.selectedTextStyle}
-                        inputSearchStyle={styles.inputSearchStyle}
-                        mode='modal'
-                        data={filtresDisponibles}
-                        labelField="typeLibelle" valueField="id"
-                        placeholder={'Selectionnez un ou plusieurs filtre'}
-                        search={true}
-                        searchPlaceholder={'Rechercher un filtre'}
-                        searchQuery={(keyword: string, labelValue: string) => {
-                            return (labelValue.match(new RegExp(keyword, 'i'))) ? true : false;}
-                        }
-                        value={selectedFiltres.map(filtre => filtre.id)}
-                        onChange={idsSelectedfiltres => updateSelectedFilters(idsSelectedfiltres, filtresDisponibles, setSelectedFiltres)}
-                        renderLeftIcon={() => (<Image source={require('@/assets/icons/filter.png')} style={styles.icon} />)}
-                        renderItem={renderFilterItem}
-                        renderSelectedItem={renderSelectedItem}
-                    />
-                </View>
-            </ThemedView>
-    )};
 
     /**
      * Affiche un panneau contenant une liste de vêtements.
@@ -128,39 +74,6 @@ export const DressingListComponent: React.FC<DressingComponentProps> = ({ veteme
     }
 
 
-    /**
-     * Rendu d'un élément de filtre dans la liste de dressing.
-     *
-     * @param {DressingListFiltreModel} item - L'élément de filtre à afficher.
-     * @returns {JSX.Element} - Un composant View contenant les informations de l'élément de filtre.
-     */
-    const renderFilterItem = (item: DressingListFiltreModel) => {
-        return (
-            <View style={styles.listItemStyle}>
-                <ThemedText type="subtitle" style={{ fontWeight: '300', fontSize: 14, fontStyle: 'italic'}}>  {item.type}</ThemedText>
-                <ThemedText type="subtitle" style={{ fontWeight: "normal" }}>{item.libelle}</ThemedText>
-            </View>
-        );
-    };
-
-    /**
-     * Rend un élément sélectionné de la liste de dressing.
-     *
-     * @param {DressingListFiltreModel} item - L'élément de la liste de dressing à afficher.
-     * @param {(item: DressingListFiltreModel) => void} unSelect - Fonction de rappel pour désélectionner l'élément.
-     * @returns {JSX.Element} - Composant Pressable affichant l'élément sélectionné avec une icône pour le désélectionner.
-     */
-    const renderSelectedItem = (item: DressingListFiltreModel, unSelect?: (item: DressingListFiltreModel) => void) => {
-        return (
-            <Pressable onPress={() => unSelect?.(item)}>
-                <View style={styles.selectedStyle}>
-                    <ThemedText type="italic" style={{fontSize:11, top:2}}> {item.type} : </ThemedText>
-                    <ThemedText type="default" style={{top:1}}>{item.libelle} </ThemedText>
-                    <Ionicons style={styles.icon} color={'white'} name="close-circle-outline" size={18} />
-                </View>
-            </Pressable>
-        )};
-
 
     return (
         <>
@@ -170,8 +83,9 @@ export const DressingListComponent: React.FC<DressingComponentProps> = ({ veteme
                     <Ionicons size={28} name="add-outline" color={Colors.dark.text} />
                 </Pressable>
             </ThemedView>
-            {renderFiltersBar()}
 
+            <DressingFiltreComponent vetementsInDressing={vetementsInDressing} setVetementsAffiches={setVetementsAffiches}/>
+            
             {showPanelGroupeVetements(groupeVetementByType(vetementsAffiches))}
         </>
     );
